+++ Editor/Tasks/CalculateAssetDependencyData.cs
@@ -467,6 +467,27 @@ namespace UnityEditor.Build.Pipeline.Tasks
             return ReturnCode.Success;
         }
 
+        private static bool IsPackedSprite(IEnumerable<ObjectIdentifier> includedObjects, ObjectIdentifier sourceTexture, TaskInput input)
+        {
+            if (EditorSettings.spritePackerMode == SpritePackerMode.Disabled)
+                return false;
+
+            foreach (var obj in includedObjects)
+            {
+                if (BuildCacheUtility.GetMainTypeForObject(obj) != typeof(Sprite))
+                    continue;
+
+                foreach (var r in ContentBuildInterface.GetPlayerDependenciesForObject(obj, input.Target, input.TypeDB, DependencyType.ValidReferences))
+                {
+                    // packed sprites never reference the source texture
+                    if (r == sourceTexture)
+                        return false;
+                }
+            }
+
+            return true;
+        }
+
         private static AssetOutput ProcessAsset(GUID asset, string assetPath, TaskInput input, HashSet<GUID> explicitAssets,
             in Dictionary<GUID, AssetOutput> implicitAssetsOutput, HashSet<GUID> packedSprites, TextureImporter spriteImporter)
         {
@@ -504,21 +525,8 @@ namespace UnityEditor.Build.Pipeline.Tasks
             if (isSprite)
             {
                 assetResult.spriteData = new SpriteImporterData();
-                assetResult.spriteData.PackedSprite = false;
                 assetResult.spriteData.SourceTexture = includedObjects.FirstOrDefault();
-
-                if (EditorSettings.spritePackerMode != SpritePackerMode.Disabled)
-                {
-                    foreach (var obj in referencedObjects)
-                    {
-                        var t = BuildCacheUtility.GetMainTypeForObject(obj);
-                        if (t != typeof(Texture2D))
-                        {
-                            assetResult.spriteData.PackedSprite = true;
-                            break;
-                        }
-                    }
-                }
+                assetResult.spriteData.PackedSprite = IsPackedSprite(includedObjects, assetResult.spriteData.SourceTexture, input);
 
 #if !UNITY_2020_1_OR_NEWER
                 if (EditorSettings.spritePackerMode == SpritePackerMode.AlwaysOn || EditorSettings.spritePackerMode == SpritePackerMode.BuildTimeOnly)