+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [2.4.1] - 2025-05-12
+- Fixed issue where sprite Atlas using late binding had incorrect dependencies.
+- Clearer error message when calling SBP with duplicate addresses.
 
 ## [2.4.0] - 2025-05-12
 - Avoid recursive dependencies in ContentFiles